# clojure: limit jvm warmup phase