this one limits the warmup phase to 5000 posts, decreasing total run time

as an interesting consequence it also slightly speeds up larger data processing (at least on my machine)