@jinyus 2024-01-19
Seeing the same speed up...I will try that on the other vm langs.

Main:
```
| Clojure | 96.90 ms | 1.08 s | 8.89 s | 10.06 s |
```

This PR:
```yaml
Clojure:

        Processing time (w/o IO): 97ms
        total: 0.88s memory: 121560k
        Processing time (w/o IO): 98ms
        total: 0.88s memory: 121332k
        Processing time (w/o IO): 98ms
        total: 0.87s memory: 125284k
        Processing time (w/o IO): 99ms
        total: 0.86s memory: 123024k
        Processing time (w/o IO): 96ms
        total: 0.87s memory: 123204k
        Processing time (w/o IO): 101ms
        total: 0.88s memory: 125436k
        Processing time (w/o IO): 99ms
        total: 0.88s memory: 124372k
        Processing time (w/o IO): 96ms
        total: 0.88s memory: 128368k
        Processing time (w/o IO): 98ms
        total: 0.86s memory: 123872k
        Processing time (w/o IO): 98ms
        total: 0.86s memory: 122432k

Clojure:

        Processing time (w/o IO): 1106ms
        total: 2.05s memory: 188796k
        Processing time (w/o IO): 1097ms
        total: 2.06s memory: 182020k
        Processing time (w/o IO): 1083ms
        total: 2.05s memory: 190136k

Clojure:

        Processing time (w/o IO): 9262ms
        total: 10.59s memory: 354256k
        Processing time (w/o IO): 9050ms
        total: 10.31s memory: 331320k
        Processing time (w/o IO): 9054ms
        total: 10.34s memory: 369572k
```