@leetwinski 2024-01-19
thanks!
that must be because of the uncertain nature of jvm's jit.. It can vary, i guess. 
Just wanted to decrease total time for large collections,.