@jinyus 2024-01-19
Running main now shows ~9s so it's not far off.