@leetwinski 2024-01-19
i was addressing this one

(this one is from yesterday's pr)
```
rocessing time (w/o IO): 8871ms
        total: 19.02s memory: 332340k
        Processing time (w/o IO): 9052ms
        total: 19.31s memory: 335032k
        Processing time (w/o IO): 8736ms
        total: 18.63s memory: 338596k
```

so, the total here is 18-19s , and now 10s, 
because i removed insanely long warmup phase for huge collections.
thought that waiting 10 extra seconds (unneeded) for every execution was annoying
))